Add input file 03.txt for AdventOfCode
This commit adds a new input file, 03.txt, to the Inputs directory for the AdventOfCode project. The contents of the file seem formatted for a specific challenge or puzzle and may include various operations or commands. This file will be used as part of the code execution and problem-solving process.